Move session persistence and transcript primitives to lib/store/
Verbatim extraction of the .rundock/ persistence octet (conversations
with the done->archived migration, lists, state, rundockDir) to
lib/store/persistence.js, and the transcript layer (convoTranscripts
cache, load/save, corrupt-file salvage, buildToolSummary) to
lib/store/transcripts.js. The only body change: workspace-root reads go
through lib/config's getWorkspace() at use time, since the root is live
state that changes on workspace switch.

lib/store/transcripts.js OWNS convoTranscripts; server.js re-exports the
module's own Map through _internal, so tests keep mutating the live
object. appendTranscript and formatTranscript stay in the composition
root: they weave the transcript primitives together with the signal
layer and the live search reconcile, which are later extraction slices.
